[TESTS] Sabrelite: Change physical memory locations
If the kernel running on the sabrelite is placed at physical address
0xC000_0000, the memblock allocator will fail. The memblock range is
calculated as follows:

   PHYS_OFFSET + arm_dma_zone_size() - 1

Since PHYS_OFFSET was 0xC000_0000 and given that arm_dma_zone_size()
returns 0x7000_0000, the result overflowed 32 bits, causing the
memblock range to be smaller than it actually is... leading to
various failures.

U-Boot places the kernel at 0x1000_0000. This scheme has been followed.
The DTS describing the emulated hardware had to be updated (artifacts
from the Vexpress-a9) according the the sabrelite manual to fit these
new memory requirements (to avoid memory regions to overlap).