[EMULATORS] Fix TX for IMX serial
When the serial console is accessed through userland in Linux, it
behaves differently than within the kernel (printk). It waits for the TX
fifo to be empty (waiting for an interrupt) and then it transmits.
After the transmission is finished, the receiver must inform the driver
that it is available again and can receive new data.