[CORE] Add generic debugging interface to emulators
This patch provides through CONFIG_DEBUG_EMU a debugging
interface that can be applied to any emulator with a
device tree node granularity.

In the device tree that describes the interface between the
hypervisor and a guest, one can now insert the "debug" attribute,
which consists in a bitfield, each bit being a flag to enable/disable
a given debugging information. The list of these flags is available
in core/include/vmm_devemu_debug.h.

Example:
	node {
		/* ... */
		debug = <0x1f>;
		/* Equivalent to:
		    debug = VMM_DEVEMU_DEBUG_PROBE |
		            VMM_DEVEMU_DEBUG_RESET |
		            VMM_DEVEMU_DEBUG_REMOVE |
		            VMM_DEVEMU_DEBUG_READ |
		            VMM_DEVEMU_DEBUG_WRITE;
		 */
	};
	Upon probing, reset, removal, read and write, debug
	information will be issued by guestX/node.

The "debug" attribute is considered to be VMM_DEVEMU_DEBUG_NONE
(no debug information) by default.