[TESTS] Sabrelite: fix (again) IPU address
The i.MX6q manual states in §2.2 (ARM Platform Memory Map) within
Table 2-2 that IPU1 starts at 0260_0000h. However, there is un
undocumented gap before this address zone, that seems to be required.

The right value: 0240_0000 is specified in Linux DT and the IPU driver
complains when this zone is not allocated to the guest.

This fixes a regression introduced by commit
302567b7c3ee38ea1cf7d90ee6691bdb222e73b3.