[CORE] vmm_stdio: fix handling of long long
Long long parameters (i.e. 64 bits) handling was
broken. The reason seems to be that due to a runtime
non-AAPCS compliant, 64-bits numbers handling would
fail in an impredictible manner.

The code removed seemed to be a workaround, since a
similar example is showned in AAPCS Sec. 2.3.2.3.
This is not needed anymore since a previous patch
fixed the runtime handling of 64-bits numbers.

This allows the wboxtest stdio/printf to pass.

Signed-off-by: Jean Guyomarc'h <jean.guyomarch@openwide.fr>