[CORE] vmm_scheduler: Minor fix in __vmm_scheduler_next2() for SMP host
On SMP host, the next VCPU dequeued from ready queue can be RESET,
PAUSED, or HALTED from another host CPU after it is dequeued and
before we acquire next->sched_lock.

To tackle the above situation, we check whether the next VCPU is in
READY state or not just before doing context switch. If the next VCPU
is not in READY state then we try to dequeue another next VCPU.

This is a precautionary patch to avoid any possibility of running
a VCPU not in READY state on SMP host. Till now no crash/hang has
been reported due to above described issue.

Signed-off-by: Anup Patel <anup@brainfault.org>